Refactored files and folders for better organization
