+++ client/src/components/Header.js
@@ -3,9 +3,9 @@ import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider'
 import AppBar from 'material-ui/AppBar'
-import HeaderButtons from './HeaderButtons'
-import HeaderAvatar from './HeaderAvatar'
-import HeaderUserName from './HeaderUserName'
+import HeaderButtons from './header/HeaderButtons'
+import HeaderAvatar from './header/HeaderAvatar'
+import HeaderUserName from './header/HeaderUserName'
 import NotificationsIcon from 'material-ui/svg-icons/social/notifications'
 import MessagesIcon from 'material-ui/svg-icons/communication/message'
 import {white} from 'material-ui/styles/colors'
